requirements.txt
python-dotenv==1.1.0
websockets==13.0
typer==0.15.3